Starting to split up Render()
