+++ field_decl.go
@@ -1,6 +1,9 @@
 package main
 
-import "strings"
+import (
+	"strings"
+	"fmt"
+)
 
 type FieldDecl struct {
 	Address    string
@@ -31,4 +34,26 @@ func parseFieldDecl(line string) *FieldDecl {
 		Referenced: len(groups["referenced"]) > 0,
 		Children:   []interface{}{},
 	}
+}
+
+func (n *FieldDecl) Render() []string {
+	fieldType := resolveType(n.Type)
+	name := strings.Replace(n.Name, "used", "", -1)
+
+	// Go does not allow the name of a variable to be called "type". For the
+	// moment I will rename this to avoid the error.
+	if name == "type" {
+		name = "type_"
+	}
+
+	suffix := ""
+	if len(n.Children) > 0 {
+		suffix = fmt.Sprintf(" = %s", renderExpression(n.Children[0])[0])
+	}
+
+	if suffix == " = (0)" {
+		suffix = " = nil"
+	}
+
+	return []string{fmt.Sprintf("%s %s%s", name, fieldType, suffix), "unknown3"}
 }
