+++ util.go
@@ -1,6 +1,9 @@
 package main
 
-import "strings"
+import (
+	"strings"
+	"strconv"
+)
 
 func ucfirst(word string) string {
 	return strings.ToUpper(string(word[0])) + word[1:]
@@ -18,4 +21,50 @@ func inStrings(item string, items []string) bool {
 	}
 
 	return false
+}
+
+func atoi(s string) int {
+	i, err := strconv.Atoi(s)
+	if err != nil {
+		panic(err)
+	}
+
+	return i
+}
+
+func removeQuotes(s string) string {
+	s = strings.TrimSpace(s)
+
+	if s == `""` {
+		return ""
+	}
+	if s == `''` {
+		return ""
+	}
+
+	if len(s) >= 2 && s[0] == '"' && s[len(s)-1] == '"' {
+		return s[1 : len(s)-2]
+	}
+	if len(s) >= 2 && s[0] == '\'' && s[len(s)-1] == '\'' {
+		return s[1 : len(s)-1]
+	}
+
+	return s
+}
+
+func atof(s string) float64 {
+	f, err := strconv.ParseFloat(s, 64)
+	if err != nil {
+		panic(err)
+	}
+
+	return f
+}
+
+func unescapeString(s string) string {
+	s = strings.Replace(s, "\\n", "\n", -1)
+	s = strings.Replace(s, "\\r", "\r", -1)
+	s = strings.Replace(s, "\\t", "\t", -1)
+
+	return s
 }
